refactor: rename shared operation item component and update imports; improve code organization
